requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
google-genai==0.3.0
python-multipart==0.0.20
pydantic==2.12.5
python-dotenv==1.2.1
audioop-lts==0.2.1


